Se modifico ubicacion del boton de registro y se agrego funcionalidad a login
